+++ src/main/java/novamachina/novacore/data/recipes/AbstractRecipeProvider.java
@@ -2,22 +2,19 @@ package novamachina.novacore.data.recipes;
 
 import java.util.Collections;
 import java.util.List;
-import java.util.concurrent.CompletableFuture;
 import net.minecraft.core.HolderLookup;
-import net.minecraft.data.PackOutput;
 import net.minecraft.data.recipes.RecipeOutput;
 import net.minecraft.data.recipes.RecipeProvider;
 
 public abstract class AbstractRecipeProvider extends RecipeProvider {
 
-  protected AbstractRecipeProvider(
-      PackOutput packOutput, CompletableFuture<HolderLookup.Provider> provider) {
-    super(packOutput, provider);
+  protected AbstractRecipeProvider(HolderLookup.Provider provider, RecipeOutput recipeOutput) {
+    super(provider, recipeOutput);
   }
 
   @Override
-  protected void buildRecipes(RecipeOutput recipeOutput) {
-    getSubProviders().forEach(subProvider -> subProvider.addRecipes(recipeOutput));
+  protected void buildRecipes() {
+    getSubProviders().forEach(subProvider -> subProvider.addRecipes(registries, this.output));
   }
 
   protected List<ISubRecipeProvider> getSubProviders() {