+++ src/main/java/novamachina/novacore/data/recipes/RecipeBuilder.java
@@ -5,7 +5,8 @@ import net.minecraft.advancements.AdvancementRequirements;
 import net.minecraft.advancements.AdvancementRewards;
 import net.minecraft.advancements.critereon.RecipeUnlockedTrigger;
 import net.minecraft.data.recipes.RecipeOutput;
-import net.minecraft.resources.ResourceLocation;
+import net.minecraft.resources.ResourceKey;
+import net.minecraft.world.item.crafting.Recipe;
 import net.minecraft.world.item.crafting.RecipeSerializer;
 import novamachina.novacore.world.item.crafting.AbstractRecipe;
 
@@ -17,18 +18,18 @@ public abstract class RecipeBuilder<T extends AbstractRecipe> {
     this.serializer = serializer;
   }
 
-  public void build(RecipeOutput consumer, ResourceLocation id) {
-    validate(id);
+  public void build(RecipeOutput consumer, ResourceKey<Recipe<?>> key) {
+    validate(key);
     Advancement.Builder advancementBuilder =
         consumer
             .advancement()
-            .addCriterion("has_the_recipe", RecipeUnlockedTrigger.unlocked(id))
-            .rewards(AdvancementRewards.Builder.recipe(id))
+            .addCriterion("has_the_recipe", RecipeUnlockedTrigger.unlocked(key))
+            .rewards(AdvancementRewards.Builder.recipe(key))
             .requirements(AdvancementRequirements.Strategy.OR);
-    consumer.accept(id, getRecipe(id), advancementBuilder.build(id));
+    consumer.accept(key, getRecipe(key), advancementBuilder.build(key.location()));
   }
 
-  protected abstract T getRecipe(ResourceLocation id);
+  protected abstract T getRecipe(ResourceKey<Recipe<?>> id);
 
-  protected void validate(ResourceLocation id) {}
+  protected void validate(ResourceKey<Recipe<?>> id) {}
 }