+++ src/main/java/novamachina/novacore/world/item/crafting/AbstractRecipe.java
@@ -2,7 +2,10 @@ package novamachina.novacore.world.item.crafting;
 
 import net.minecraft.core.HolderLookup;
 import net.minecraft.world.item.ItemStack;
+import net.minecraft.world.item.crafting.PlacementInfo;
 import net.minecraft.world.item.crafting.Recipe;
+import net.minecraft.world.item.crafting.RecipeBookCategories;
+import net.minecraft.world.item.crafting.RecipeBookCategory;
 import net.minecraft.world.item.crafting.RecipeInput;
 import net.minecraft.world.level.Level;
 
@@ -19,12 +22,12 @@ public abstract class AbstractRecipe implements Recipe<RecipeInput> {
   }
 
   @Override
-  public boolean canCraftInDimensions(int pWidth, int pHeight) {
-    return true;
+  public PlacementInfo placementInfo() {
+    return PlacementInfo.NOT_PLACEABLE;
   }
 
   @Override
-  public ItemStack getResultItem(HolderLookup.Provider provider) {
-    return ItemStack.EMPTY;
+  public RecipeBookCategory recipeBookCategory() {
+    return RecipeBookCategories.CRAFTING_MISC;
   }
 }