+++ CHANGELOG.md
@@ -4,6 +4,25 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/), and this project adheres to [NeoForged Semantic Versioning](https://docs.neoforged.net/docs/gettingstarted/versioning).
 
+## [4.0.0](https://github.com/NovaMachina-Mods/NovaCore/compare/v3.0.0...v4.0.0) - 2025-01-15
+
+### Added
+- Registry key to `AbstractRegistry`
+- `BlockBehavior.Properties` to block factory methods
+- `Item.Properties` to item factory methods
+- `BucketItem` factory method
+- Add tag to tag
+
+### Changed
+- Update to 1.21.4
+- Require NeoForge 21.4.20-beta
+-
+### Removed
+- Burnable block factory method
+- Conditions from `LootModifierDefinition`
+- `IngredientUtils` is essentially defunct
+- Fluid textures are no longer set in `FluidDefinition`
+
 ## [3.0.0](https://github.com/NovaMachina-Mods/NovaCore/compare/v2.0.0...v3.0.0) - 2024-07-09
 
 ### Added